+++ baekjoon-practice/src/_2023112/_08_10988.java
@@ -1,4 +1,4 @@
-package _202311;
+package _2023112;
 
 import java.io.BufferedReader;
 import java.io.IOException;
