+++ baekjoon-practice/src/_2023112/_09_25083.java
@@ -1,4 +1,4 @@
-package _202311;
+package _2023112;
 
 /**
  * 25083 새싹
