+++ baekjoon-practice/src/_2023112/_1008.java
@@ -1,12 +1,12 @@
-package _202311;
+package _2023112;
 
 import java.util.Scanner;
 
 /**
- * 10430 나머지
- * https://www.acmicpc.net/problem/10430
+ * https://www.acmicpc.net/problem/1008
+ * 1008 A/B
  */
-public class _10430 {
+public class _1008 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         int A = scanner.nextInt();
