+++ baekjoon-practice/src/_2023112/_17_1193.java
@@ -1,4 +1,4 @@
-package _202311;
+package _2023112;
 
 import java.util.Scanner;
