+++ baekjoon-practice/src/_2023112/_2557.java
@@ -1,4 +1,4 @@
-package _202311;
+package _2023112;
 
 /**
  * 2557 Hello World
